+++ Assets/javascript/ppfeufer.js
@@ -1,12 +1,60 @@
 /* global Masonry */
 
-// jQuery(document).ready(($) => {
-jQuery(document).ready(() => {
+jQuery(document).ready(($) => {
     'use strict';
 
-    const copyButton = {
-        copyCode: `<svg style="width: 16px; height: 16px;"><use href="#copy-code"></use></svg>`,
-        codeCopied: `<svg style="width: 16px; height: 16px;"><use href="#code-copied"></use></svg>`
+    // HTML elements used in the script
+    const elements = {
+        copyButton: {
+            copyCode: '<svg style="width: 16px; height: 16px;"><use href="#copy-code"></use></svg>',
+            codeCopied: '<svg style="width: 16px; height: 16px;"><use href="#code-copied"></use></svg>'
+        }
+    };
+
+    /**
+     * Extend links to external website.
+     *
+     * » add target="_blank"
+     * » add referrerpolicy="no-referrer"
+     * » add rel="noopener noreferrer"
+     */
+    const externalLinks = () => {
+        // Get the current location hostname
+        const internalHost = [location.hostname];
+
+        // Regex pattern to match HTTP and HTTPS
+        const protocolPattern = /^https?:\/\//i;
+
+        // Walk through all links on the current page.
+        $('a').each((index, element) => { // jshint ignore:line
+            // Get the href attribute of the link
+            const href = $(element).attr('href');
+
+            // Check if it's an HTTP link
+            if (protocolPattern.test(href)) {
+                // Get the hostname of the link
+                const hrefHostname = new URL(href).hostname;
+
+                // Check if the hostname is not in the internalHost array or if the link has the class 'external-link',
+                // and add the target and classes and attributes to the link element.
+                if (
+                    // Check if the hostname is not in the internalHost array
+                    $.inArray(hrefHostname, internalHost) === -1
+                    // Check if the link has the class 'external-link'
+                    || $(element).hasClass('external-link') // jshint ignore:line
+                    // Check if the parent <li> element has the class 'external-link'.
+                    // This is useful for links in navigation menus that are marked as
+                    // external links, as WordPress does not add the 'external-link'
+                    // class to the <a> element, but to the parent <li> element.
+                    || $(element).parent('li.menu-item').hasClass('external-link') // jshint ignore:line
+                ) {
+                    $(element).addClass('external-link');
+                    $(element).attr('target', '_blank');
+                    $(element).attr('rel', 'noopener noreferrer');
+                    $(element).attr('referrerpolicy', 'no-referrer');
+                }
+            }
+        });
     };
 
     /**
@@ -23,10 +71,10 @@ jQuery(document).ready(() => {
             await navigator.clipboard.writeText(code.innerText);
 
             // Visual feedback
-            button.innerHTML = copyButton.codeCopied;
+            button.innerHTML = elements.copyButton.codeCopied;
 
             setTimeout(() => {
-                button.innerHTML = copyButton.copyCode;
+                button.innerHTML = elements.copyButton.copyCode;
             }, 5000);
         } catch (err) {
             console.error('Failed to copy code:', err);
@@ -47,7 +95,7 @@ jQuery(document).ready(() => {
         blocks.forEach((block) => {
             const button = document.createElement('span');
 
-            button.innerHTML = copyButton.copyCode;
+            button.innerHTML = elements.copyButton.copyCode;
             button.classList.add('copy-to-clipboard');
             button.addEventListener('click', () => {
                 copyCode(block, button);
@@ -112,6 +160,7 @@ jQuery(document).ready(() => {
 
     // Use setTimeout instead of custom sleep function
     // setTimeout(addCopyButtons, 2000);
+    externalLinks();
     addCopyButtons();
     initMasonry();
     // initStickyElements();